slidedeck_transform/requirements.txt
python-pptx
pillow
regex
pytest
pep8
autopep8
